Add chown to ceph user for mon directory.
The actual user-id in this article is very vague. It starts with having you create "a user" that can ssh and sudo, then creates lots of stuff as root, then starts ceph-mon as .. some random uid which will not necessarily be able to read root-owned files, and is not certain to be "ceph" either.

While this manual perhaps gets you something that runs, you would probably be using systemctl or whatever upstart handler your OS comes with at step #6, and not have the mon hang off your login-as-some-user ssh session. At least systemd unit files will set user and group to "ceph" for the ceph-mon, so if you mkfs'd and ran ceph-mon as something else before, then this is when you will notice if perms are bad.